requirements.txt
flask
flask-sqlalchemy
pyjwt
werkzeug
psycopg2-binary
gunicorn